Move recursive matching out to isolated module
This is not really required at the top level of the package.

Next step is going to be adding something to go and complete the graph
from the search index. This is the other functionality of the package.

This way the two main functions of the package are implemented at the
top level, and these alone.