Implement Complete & Contains matching
Nodes and Edges can now be matched at different levels. Previously a
complete match was required, e.g.

VB != VBD

Now there are two options, Complete and Contains. Complete is as it was,
Contains is:

VB == VBD

In the future, regex matchers may be required. These would be much
slower so I'm seeing how we can make do for now.